+++ CHANGELOG.md
@@ -21,9 +21,10 @@ k.vec2();
 - disabled context menu on canvas
 - prevented default behavior for 'tab' and function keys
 - removed `loadAseprite()` and made it an external plugin under `plugins/aseprite.js`
-- supports newer version of pedit
 - added a plugin system, see the `multiboom` example and `src/plugins`
+- (**BREAK**) removed support for `.kbmsprite`, supports newer version of `.pedit` through pedit plugin
 - added `numFrames()` by `sprite()`
+- added `screenShot()` that returns of a png base64 data url for a screenshot
 
 # v0.3.0 "King Dedede...Bug!"
 - (**BREAK**) removed `pause()` and `paused()` in favor to `kaboom.debug.paused`